refactor proxyserver
