+++ internal/proxyserver/state_test.go
@@ -5,6 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/safedep/pmg/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -60,4 +61,17 @@ func TestIsRunningDeadPID(t *testing.T) {
 
 func TestStateFilePath(t *testing.T) {
 	assert.Equal(t, "/some/dir/proxy-state.json", stateFilePath("/some/dir"))
+}
+
+func TestResolveStatePath(t *testing.T) {
+	cfg := config.Get()
+
+	t.Run("flag override wins", func(t *testing.T) {
+		assert.Equal(t, "/custom/proxy.json", ResolveStatePath("/custom/proxy.json", cfg))
+	})
+
+	t.Run("defaults to cacheDir", func(t *testing.T) {
+		want := filepath.Join(cfg.CacheDir(), "proxy-state.json")
+		assert.Equal(t, want, ResolveStatePath("", cfg))
+	})
 }
